% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bushfire_app.R
\name{run_bushfire_app}
\alias{run_bushfire_app}
\title{Launch the Bushfire Shiny App}
\usage{
run_bushfire_app()
}
\value{
Runs the Shiny app in your default browser.
}
\description{
Opens the interactive bushfire risk exploration app.
}
